Abstract post type query handling
