+++ api/wordpress/_global/getPostTypeById.js
@@ -1,16 +1,13 @@
 import queryPostById from '../posts/queryPostById'
-import {initializeWpApollo, createWpApolloClient} from '../connector'
 import queryPageById from '../pages/queryPageById'
 import {isHierarchicalPostType} from './postTypes'
-import formatBlockData from '@/functions/formatBlockData'
 import queryEventById from '../events/queryEventById'
 import queryCareerById from '../careers/queryCareerById'
 import queryServiceById from '../services/queryServiceById'
 import queryTeamById from '../teams/queryTeamById'
 import queryPortfolioById from '../portfolios/queryPortfolioById'
 import queryTestimonialById from '../testimonials/queryTestimonialById'
-import formatDefaultSeoData from '@/functions/formatDefaultSeoData'
-import getMenus from '../menus/getMenus'
+import processPostTypeQuery from './processPostTypeQuery'
 
 /**
  * Retrieve single post by specified identifier.
@@ -49,77 +46,5 @@ export default async function getPostTypeById(
   // Retrieve post type query.
   const query = postTypeQuery?.[postType] ?? null
 
-  // Get/create Apollo instance.
-  const apolloClient = preview
-    ? createWpApolloClient(true)
-    : initializeWpApollo()
-
-  // Set up return object.
-  const response = {
-    apolloClient,
-    error: false,
-    errorMessage: null
-  }
-
-  // If no query is set for given post type, return error message.
-  if (!query) {
-    return {
-      apolloClient,
-      error: true,
-      errorMessage: `Post type \`${postType}\` is not supported.`
-    }
-  }
-
-  // Execute query.
-  response.post = await apolloClient
-    .query({query, variables: {id, idType}})
-    .then((res) => {
-      const {homepageSettings, siteSeo, menus, ...postData} = res.data
-
-      // Retrieve menus.
-      response.menus = getMenus(menus)
-
-      // Retrieve default SEO data.
-      response.defaultSeo = formatDefaultSeoData({homepageSettings, siteSeo})
-
-      const post = postData?.[postType]
-
-      // Set error props if data not found.
-      if (!post) {
-        response.error = true
-        response.errorMessage = `An error occurred while trying to retrieve data for ${postType} "${id}."`
-
-        return null
-      }
-
-      return post
-    })
-    .then(async (post) => {
-      // Add slug/ID to post.
-      const newPost = {
-        ...post,
-        slug: id
-      }
-
-      if ('basic' === preview || !post || !post?.blocksJSON) {
-        return post
-      }
-
-      // Handle blocks.
-      newPost.blocks = await formatBlockData(
-        JSON.parse(newPost.blocksJSON) ?? []
-      )
-
-      delete newPost.blocksJSON
-
-      return newPost
-    })
-    .catch((error) => {
-      response.error = true
-      response.errorMessage = error.message
-
-      return null
-    })
-
-  return response
+  return processPostTypeQuery(postType, id, query, {id, idType}, preview)
 }
